backend/requirements.txt
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
ImageIO==2.37.2
imbalanced-learn==0.14.1
imblearn==0.0
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
lime==0.2.0.1
llvmlite==0.46.0
matplotlib==3.10.8
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
packaging==26.0
pandas==3.0.0
pillow==12.1.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
shap==0.50.0
six==1.17.0
sklearn-compat==0.1.5
slicer==0.0.8
threadpoolctl==3.6.0
tifffile==2026.1.28
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3

fastapi
uvicorn[standard]
python-dotenv
supabase
pydantic

